requirements.txt
django
gunicorn
telebot
vk_api
python-dotenv
requests
openpyxl==.1.5